created relay and timer class, created logic in idle state  can context set to error

+++ bathroomfan/calcbehaviour.h
@@ -15,7 +15,7 @@ public:
         CALULATE_DEFAULT
     } calc_e;
 
-    virtual int calculate(calc_e *enumValue, int *value);
+    virtual double calculate(calc_e *enumValue, double *value);
 };
 
 #endif // CALCBEHAVIOUR_H
